refactor: centralize _LazyModule in hera/utils/lazy.py, remove 4 duplicate class definitions refs #935
