+++ hera/measurements/GIS/utils.py
@@ -1,30 +1,9 @@
 import math  # stdlib, cheap
+from hera.utils.lazy import _LazyModule
 
 # Heavy libraries are loaded on first attribute access so that
 # `from hera.measurements.GIS.utils import WSG84, ITM, convertCRS`
 # is instant even when geopandas / xarray are not yet imported.
-class _LazyModule:
-    """Proxy that imports a module the first time any attribute is accessed."""
-    __slots__ = ("_name", "_mod")
-
-    def __init__(self, name):
-        object.__setattr__(self, "_name", name)
-        object.__setattr__(self, "_mod", None)
-
-    def _load(self):
-        import importlib
-        mod = importlib.import_module(object.__getattribute__(self, "_name"))
-        object.__setattr__(self, "_mod", mod)
-        return mod
-
-    def __getattr__(self, item):
-        mod = object.__getattribute__(self, "_mod") or self._load()
-        return getattr(mod, item)
-
-    def __call__(self, *a, **kw):
-        mod = object.__getattribute__(self, "_mod") or self._load()
-        return mod(*a, **kw)
-
 
 numpy    = _LazyModule("numpy")
 np       = numpy
