+++ hera/measurements/experiment/analysis.py
@@ -1,18 +1,5 @@
 from hera.datalayer import datatypes
-
-class _LazyModule:
-    __slots__ = ("_name", "_mod")
-    def __init__(self, name):
-        object.__setattr__(self, "_name", name)
-        object.__setattr__(self, "_mod", None)
-    def _load(self):
-        import importlib
-        mod = importlib.import_module(object.__getattribute__(self, "_name"))
-        object.__setattr__(self, "_mod", mod)
-        return mod
-    def __getattr__(self, item):
-        mod = object.__getattribute__(self, "_mod") or self._load()
-        return getattr(mod, item)
+from hera.utils.lazy import _LazyModule
 
 pandas = _LazyModule("pandas")
